refactor(todo): modularize scaffold layout
