+++ toondo/packages/presentation/lib/navigation/router.dart
@@ -9,7 +9,7 @@ import 'package:presentation/views/mypage/account_setting/phone_number_change_sc
 import 'package:presentation/views/mypage/display_setting/display_setting_screen.dart';
 import 'package:presentation/views/mypage/help_guide/help_guide_screen.dart';
 import 'package:presentation/views/mypage/notification_setting/notification_setting_screen.dart';
-import 'package:presentation/views/todo/todo_manage_view.dart';
+import 'package:presentation/views/todo/todo_manage_screen.dart';
 import 'package:provider/provider.dart';
 import 'package:presentation/views/home/home_screen.dart';
 import 'package:presentation/views/goal/goal_progress_screen.dart';
@@ -43,7 +43,7 @@ class AppRouter {
       case RoutePaths.goalAnalisys:
         return MaterialPageRoute(builder: (_) => GoalProgressScreen());
       case RoutePaths.todoManage:
-        return MaterialPageRoute(builder: (_) => TodoManageView());
+        return MaterialPageRoute(builder: (_) => TodoManageScreen());
       case RoutePaths.mypage:
         return MaterialPageRoute(builder: (_) => MyPageScreen());
       case RoutePaths.notificationSetting:
