+++ toondo/packages/presentation/lib/widgets/navigation/bottom_navigation_bar_widget.dart
@@ -4,7 +4,7 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:presentation/views/goal/goal_manage_view.dart';
 import 'package:presentation/views/goal/goal_progress_screen.dart';
 import 'package:presentation/views/mypage/my_page_screen.dart';
-import 'package:presentation/views/todo/todo_manage_view.dart';
+import 'package:presentation/views/todo/todo_manage_screen.dart';
 
 class BottomNavigationBarWidget extends StatelessWidget {
   const BottomNavigationBarWidget({Key? key}) : super(key: key);
@@ -36,7 +36,7 @@ class BottomNavigationBarWidget extends StatelessWidget {
               onTap: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(builder: (_) => TodoManageView()),
+                  MaterialPageRoute(builder: (_) => TodoManageScreen()),
                 );
               },
             ),
